move match sorter into utils folder
